requirements.txt
asgiref==3.5.0
Django==4.0.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
djongo==1.3.6
Pillow==9.0.1
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
